+++ ChurnR.Core/CutoffProcessor/MinimalCutoffProcessor.cs
@@ -1,4 +1,4 @@
-﻿namespace ChurnR.Core.Processors.Cutoff;
+﻿namespace ChurnR.Core.CutoffProcessor;
 
 public class MinimalCutoffProcessor(int minimum) : IProcessor
 {
