+++ ChurnR.Core/Reporter/ChartJsReporter.cs
@@ -1,6 +1,6 @@
-﻿using ChurnR.Core.Processors;
+﻿using ChurnR.Core.CutoffProcessor;
 
-namespace ChurnR.Core.Reporters;
+namespace ChurnR.Core.Reporter;
 
 public class ChartJsReporter(TextWriter output, IProcessor cutOffProcessor) : BaseReporter(output, cutOffProcessor)
 {
