+++ ChurnR.Core/Reporter/CsvReporter.cs
@@ -1,7 +1,7 @@
 ﻿using System.Text;
-using ChurnR.Core.Processors;
+using ChurnR.Core.CutoffProcessor;
 
-namespace ChurnR.Core.Reporters;
+namespace ChurnR.Core.Reporter;
 
 public class CsvReporter(TextWriter output, IProcessor cutOffProcessor) : BaseReporter(output, cutOffProcessor)
 {
